Refactor copyTemplateFile to utility
